013+++ CRUD постов (часть 1, показать-создать пост)

+++ app/Models/Post.php
@@ -2,10 +2,12 @@
 
 namespace App\Models;
 
+use Carbon\Carbon;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Cviebrock\EloquentSluggable\Sluggable;
+use Illuminate\Support\Str;
 
 class Post extends Model
 {
@@ -15,17 +17,17 @@ class Post extends Model
     const IS_DRAFT = 0;
     const IS_PUBLIC = 1;
 
-    protected $fillable = ['title', 'content'];
+    protected $fillable = ['title', 'content', 'date'];
 
     public function category() 
     {
-        return $this->hasOne(Category::class);
+        return $this->belongsTo(Category::class);
     }
 
 
     public function author() 
     {
-        return $this->hasOne(User::class);
+        return $this->belongsTo(User::class, 'user_id');
     }
 
     public function tags() 
@@ -66,17 +68,25 @@ class Post extends Model
 
     public function remove()
     {
-        Storage::delete('uploads/'. $this->image);
+        $this->removeImage();
         $this->delete();
+    }
+
+    public function removeImage()
+    {
+        if($this->image != null)
+        {
+            Storage::delete('uploads/' . $this->image);
+        }
     }
 
     public function uploadImage($image)
     {
         if ($image == null) { return; }
 
-        Storage::delete('uploads/'. $this->image);
-        $filename = str_random(10) . '.' . $image->extension();
-        $image->saveAs('uploads', $filename);
+        $this->removeImage();
+        $filename = Str::random(10) . '.' . $image->extension();
+        $image->storeAs('uploads', $filename);
         $this->image = $filename;
         $this->save();
     }
@@ -85,10 +95,10 @@ class Post extends Model
     {
         if($this->image == null) 
         {
-            return '/img/no-image.png';
+            return '/img/no-avatar.png';
         }
         
-        return 'uploads' . $this->image;
+        return '/uploads/' . $this->image;
     }
 
 
@@ -104,7 +114,7 @@ class Post extends Model
     {
         if ($ids == null) { return; }
 
-        $this->tags->sync($ids);
+        $this->tags()->sync($ids);
     }
 
 
@@ -152,5 +162,24 @@ class Post extends Model
         
     }
 
-    
+
+    public function setDateAttribute($value)
+    {
+        $date = Carbon::createFromFormat('d/m/y', $value)->format('Y-m-d');
+        $this->attributes['date'] = $date;
+    }
+
+    public function getCategoryTitle()
+    {
+        return ($this->category != null)
+            ? $this->category->title
+            : 'Нет категории';
+    }
+
+    public function getTagsTitle()
+    {
+        return (!$this->tags->isEmpty())
+            ? implode(', ', $this->tags->pluck('title')->all())
+            : 'Нет тегов';
+    }
 }
